working on plic.

+++ src/interrupt/plic.rs
@@ -24,7 +24,7 @@ impl PLIC {
         for irq in irqs {
             // set irq priorities
             unsafe {
-                (self.address + irq as usize * size_of::<u32>() as usize).write_volatile(&1u32);
+                (self.address + irq as usize * size_of::<u32>()).write_volatile(&1u32);
             }
         }
     }
